Add optional periodic refresh of the MySQL connection
On the Aurora DB use case where autoscaling of read replicas
is available we want to be able to start using them as they come
up and not simply be sticked to the same connection all the time.
So every x seconds roll a dice and you get a 50/50 chance of killing
the worker and having it restarted thus getting a new connection.